Makes iCal file spec-compliant
(The iCal format is [RFC
5545](https://datatracker.ietf.org/doc/html/rfc5545).)

Checked with [an online validator](https://icalendar.org/validator.html).

Also fixes the import warning you get if you try to import the file into
an Exchange calendar via Outlook, which was caused by the URL field not
existing in Exchange (in violation of spec).

I'm not sure if Apple's Calendar actually uses the URL field, but in
theory it's supposed to be something like a link back to the source
of the event, rather than a video call client URL. The former contents
were moved to the general-purpose description field.

Key differences:

 - A UID per event
 - A PRODID per calendar
 - A DTSTAMP per event
 - No URL field per event